+++ oshi-core-java25/src/main/java/oshi/driver/windows/perfmon/ProcessInformationFFM.java
@@ -7,6 +7,7 @@ package oshi.driver.windows.perfmon;
 import static oshi.driver.common.windows.perfmon.PerfmonConstants.PROCESS;
 import static oshi.driver.common.windows.perfmon.PerfmonConstants.WIN32_PERFPROC_PROCESS_WHERE_IDPROCESS_0;
 import static oshi.driver.common.windows.perfmon.PerfmonConstants.WIN32_PERFPROC_PROCESS_WHERE_NAME_TOTAL;
+import static oshi.driver.common.windows.perfmon.PerfmonConstants.WIN32_PERFPROC_PROCESS_WHERE_NOT_NAME_LIKE_TOTAL;
 
 import java.util.List;
 import java.util.Map;
@@ -14,6 +15,7 @@ import java.util.Map;
 import oshi.annotation.concurrent.ThreadSafe;
 import oshi.driver.common.windows.perfmon.ProcessInformation.HandleCountProperty;
 import oshi.driver.common.windows.perfmon.ProcessInformation.IdleProcessorTimeProperty;
+import oshi.driver.common.windows.perfmon.ProcessInformation.ProcessPerformanceProperty;
 import oshi.util.platform.windows.PerfCounterQueryFFM;
 import oshi.util.platform.windows.PerfCounterWildcardQueryFFM;
 import oshi.util.tuples.Pair;
@@ -25,6 +27,16 @@ import oshi.util.tuples.Pair;
 public final class ProcessInformationFFM {
 
     private ProcessInformationFFM() {
+    }
+
+    /**
+     * Returns process counters.
+     *
+     * @return Process counters for each process.
+     */
+    public static Pair<List<String>, Map<ProcessPerformanceProperty, List<Long>>> queryProcessCounters() {
+        return PerfCounterWildcardQueryFFM.queryInstancesAndValues(ProcessPerformanceProperty.class, PROCESS,
+                WIN32_PERFPROC_PROCESS_WHERE_NOT_NAME_LIKE_TOTAL);
     }
 
     /**